Fix unit tests.
